Solution for 2020 Day 03, Challenge 02

+++ 2020/Day03/Models/Answers.cs
@@ -8,17 +8,17 @@ namespace Advent2020.Models
 {
     class Answers
     {
-        private int _answerTest1;
-        private int _answerTest2;
-        private int _answer1;
-        private int _answer2;
+        private long _answerTest1;
+        private long _answerTest2;
+        private long _answer1;
+        private long _answer2;
 
         public bool AnswerFoundTest1 { get; set; } = false;
         public bool AnswerFoundTest2 { get; set; } = false;
         public bool AnswerFound1 { get; set; } = false;
         public bool AnswerFound2{ get; set; } = false;
 
-        public int AnswerTest1 {
+        public long AnswerTest1 {
             get => this._answerTest1;
             set
             {
@@ -27,7 +27,7 @@ namespace Advent2020.Models
             }
         }
 
-        public int AnswerTest2 {
+        public long AnswerTest2 {
             get => this._answerTest2;
             set
             {
@@ -36,7 +36,7 @@ namespace Advent2020.Models
             }
         }
 
-        public int Answer1 {
+        public long Answer1 {
             get => this._answer1;
             set
             {
@@ -45,7 +45,7 @@ namespace Advent2020.Models
             }
         }
 
-        public int Answer2 {
+        public long Answer2 {
             get => this._answer2;
             set
             {
@@ -54,7 +54,7 @@ namespace Advent2020.Models
             }
         }
 
-        public Answers SetAnswer(int value, bool isTestAnswer, bool isAnswer2)
+        public Answers SetAnswer(long value, bool isTestAnswer, bool isAnswer2)
         {
             if (isTestAnswer)
             {
